Add a GTK4 variant
This commits adds a GTK4 variant, libcolord-gtk4. An effort was made
to keep the changes as separate and self-contained as possible.

The new GTK4 code of the sample window and widget is introduced in
separated files, reusing the cd-sample-{window,widget}.h headers.
There was only one place in CdWindow that required split GTK3 / GTK4
paths, and fortunately it is a rather trivial one.